decrease quantity of files in nextdate
