+++ internal/nextdate/nextdate.go
@@ -2,9 +2,6 @@ package nextdate
 
 import (
 	"time"
-
-	count_nextdate "github.com/kasariks/project_for_graduating/internal/nextdate/count"
-	validate_nextdate "github.com/kasariks/project_for_graduating/internal/nextdate/validate"
 )
 
 const DateFormat = "20060102"
@@ -15,11 +12,11 @@ func NextDate(now time.Time, dstart string, repeat string) (string, error) {
 		return "", err
 	}
 
-	if err := validate_nextdate.ValidateRepeatParam(repeat); err != nil {
+	if err := ValidateRepeatParam(repeat); err != nil {
 		return "", err
 	}
 
-	countedTime, err := count_nextdate.CountNextDate(now, date, repeat)
+	countedTime, err := CountNextDate(now, date, repeat)
 	if err != nil {
 		return "", err
 	}
